# src/main/java/br/com/becommerce/product/catalog/api/shoe/model/Shoe.java
package br.com.becommerce.product.catalog.api.shoe.model;

import javax.persistence.Entity;
import javax.persistence.GeneratedValue;
import javax.persistence.GenerationType;
import javax.persistence.Id;
import javax.validation.constraints.Max;
import javax.validation.constraints.Min;
import javax.validation.constraints.NotBlank;
import javax.validation.constraints.NotNull;
import javax.validation.constraints.PositiveOrZero;
import javax.validation.constraints.Size;

@Entity
public class Shoe {

	@Id
	@GeneratedValue(strategy = GenerationType.AUTO)
	private long id;

	@Min(10)
	@Max(50)
	private int size;

	@NotBlank
	@Size(max = 32)
	private String brandName;

	@PositiveOrZero
	private double price;

	@NotNull
	private Color color;

	public long getId() {
		return id;
	}

	public void setId(long id) {
		this.id = id;
	}

	public int getSize() {
		return size;
	}

	public void setSize(int size) {
		this.size = size;
	}

	public String getBrandName() {
		return brandName;
	}

	public void setBrandName(String brandName) {
		this.brandName = brandName;
	}

	public double getPrice() {
		return price;
	}

	public void setPrice(double price) {
		this.price = price;
	}

	public Color getColor() {
		return color;
	}

	public void setColor(Color color) {
		this.color = color;
	}
}